--- benchmarks/scenarios/efficiency.json
@@ -1,26 +1,0 @@
-[
-  {
-    "id": "ef-simple-echo",
-    "category": "efficiency",
-    "input": "Echo the message 'hello world' using the echo tool.",
-    "success_criteria": [
-      {"type": "tool_used", "tool": "echo"},
-      {"type": "tool_call_count_max", "max": 2},
-      {"type": "response_contains", "text": "hello world"}
-    ],
-    "timeout_secs": 30,
-    "max_tool_iterations": 5
-  },
-  {
-    "id": "ef-single-file-read",
-    "category": "efficiency",
-    "input": "Read /tmp/ironclaw_bench_efficiency.txt and tell me its contents.",
-    "success_criteria": [
-      {"type": "tool_used", "tool": "read_file"},
-      {"type": "tool_call_count_max", "max": 2},
-      {"type": "response_contains", "text": "efficiency test content"}
-    ],
-    "timeout_secs": 30,
-    "max_tool_iterations": 5
-  }
-]
