--- benchmarks/scenarios/error_recovery.json
@@ -1,26 +1,0 @@
-[
-  {
-    "id": "er-read-missing-file",
-    "category": "error_recovery",
-    "input": "Read the file /tmp/ironclaw_bench_nonexistent_12345.txt and tell me what happened.",
-    "success_criteria": [
-      {"type": "tool_used", "tool": "read_file"},
-      {"type": "tool_call_count_max", "max": 3},
-      {"type": "response_matches", "pattern": "(?i)(not found|does not exist|error|no such file|couldn't|unable)"}
-    ],
-    "timeout_secs": 30,
-    "max_tool_iterations": 5
-  },
-  {
-    "id": "er-invalid-json",
-    "category": "error_recovery",
-    "input": "Parse this invalid JSON: {broken and tell me what's wrong with it.",
-    "success_criteria": [
-      {"type": "tool_used", "tool": "json"},
-      {"type": "tool_call_count_max", "max": 3},
-      {"type": "response_matches", "pattern": "(?i)(invalid|error|parse|syntax|malformed)"}
-    ],
-    "timeout_secs": 30,
-    "max_tool_iterations": 5
-  }
-]
