--- benchmarks/scenarios/memory_operations.json
@@ -1,27 +1,0 @@
-[
-  {
-    "id": "mo-full-cycle",
-    "category": "memory_operations",
-    "input": "Write a note to memory at 'bench/cycle-test.md' with content 'The answer is 42'. Then list the memory tree to confirm it exists. Then read it back and tell me the answer.",
-    "success_criteria": [
-      {"type": "tool_used", "tool": "memory_write"},
-      {"type": "tool_used", "tool": "memory_tree"},
-      {"type": "tool_used", "tool": "memory_read"},
-      {"type": "response_contains", "text": "42"}
-    ],
-    "timeout_secs": 30,
-    "max_tool_iterations": 15
-  },
-  {
-    "id": "mo-search-recall",
-    "category": "memory_operations",
-    "input": "Save a note to memory at 'bench/search-test.md' with content 'Project deadline is March 15th'. Then search memory for 'deadline' and tell me when it is.",
-    "success_criteria": [
-      {"type": "tool_used", "tool": "memory_write"},
-      {"type": "tool_used", "tool": "memory_search"},
-      {"type": "response_contains", "text": "March 15"}
-    ],
-    "timeout_secs": 30,
-    "max_tool_iterations": 10
-  }
-]
