--- benchmarks/scenarios/tool_selection.json
@@ -1,37 +1,0 @@
-[
-  {
-    "id": "ts-time-query",
-    "category": "tool_selection",
-    "input": "What time is it right now?",
-    "success_criteria": [
-      {"type": "tool_used", "tool": "time"},
-      {"type": "tool_not_used", "tool": "shell"},
-      {"type": "response_matches", "pattern": "\\d{1,2}:\\d{2}"}
-    ],
-    "timeout_secs": 30,
-    "max_tool_iterations": 5
-  },
-  {
-    "id": "ts-direct-answer",
-    "category": "tool_selection",
-    "input": "What is 2 + 2?",
-    "success_criteria": [
-      {"type": "response_contains", "text": "4"},
-      {"type": "tool_call_count_max", "max": 0}
-    ],
-    "timeout_secs": 30,
-    "max_tool_iterations": 5
-  },
-  {
-    "id": "ts-echo-not-shell",
-    "category": "tool_selection",
-    "input": "Use the echo tool to say 'benchmark test'.",
-    "success_criteria": [
-      {"type": "tool_used", "tool": "echo"},
-      {"type": "tool_not_used", "tool": "shell"},
-      {"type": "response_contains", "text": "benchmark test"}
-    ],
-    "timeout_secs": 30,
-    "max_tool_iterations": 5
-  }
-]
