--- benchmarks/trajectories/setup/tool-restriction.json
@@ -1,19 +1,0 @@
-{
-    "name": "tool-restriction-echo-only",
-    "description": "Verify that tool restriction limits available tools to only echo",
-    "tags": ["setup", "tools"],
-    "setup": {
-        "tools": ["echo"]
-    },
-    "turns": [
-        {
-            "user": "Use the echo tool to say 'restricted test'",
-            "assertions": {
-                "tools_called": ["echo"],
-                "tools_not_called": ["time", "json", "http"],
-                "response_contains": ["restricted test"]
-            }
-        }
-    ],
-    "timeout_secs": 60
-}
