--- benchmarks/trajectories/tool-selection/pick-echo-tool.json
@@ -1,21 +1,0 @@
-{
-  "name": "pick-echo-tool",
-  "description": "Agent should use echo tool, not shell",
-  "tags": ["tools", "basic"],
-  "setup": {
-    "tools": ["echo", "shell"]
-  },
-  "timeout_secs": 30,
-  "max_tool_iterations": 5,
-  "turns": [
-    {
-      "user": "Use the echo tool to say 'benchmark test'.",
-      "assertions": {
-        "tools_called": ["echo"],
-        "tools_not_called": ["shell"],
-        "response_contains": ["benchmark test"],
-        "max_tool_calls": 3
-      }
-    }
-  ]
-}
