--- benchmarks/trajectories/tool-selection/pick-time-tool.json
@@ -1,20 +1,0 @@
-{
-  "name": "pick-time-tool",
-  "description": "Agent should use the time tool, not shell",
-  "tags": ["tools", "basic"],
-  "setup": {
-    "tools": ["time", "echo", "shell"]
-  },
-  "timeout_secs": 30,
-  "max_tool_iterations": 5,
-  "turns": [
-    {
-      "user": "What time is it right now?",
-      "assertions": {
-        "tools_called": ["time"],
-        "tools_not_called": ["shell"],
-        "max_tool_calls": 3
-      }
-    }
-  ]
-}
