+++ src/lib.rs
@@ -40,7 +40,6 @@
 
 pub mod agent;
 pub mod app;
-pub mod benchmark;
 pub mod boot_screen;
 pub mod bootstrap;
 pub mod channels;
